Project Update
Refactoring code from Mern Jobify project. Most of it has to go but it's acting as a good template. At this point we have a countryController that we can use to fetch country information from restcountries.com and return it to the user. We have a front end of sorts that the user can enter a country into, hit submit, and get back information about that country. When we enter 'chad,' for example, we get back our data and dump 'The population of chad is 16425859' into a paragraph on our frontend. Success!