Project Update
Went through my example-data.json and annotated. If I am going to present information, I will need to know what it means. Added to todo.md and created plan.md. Plan is more about decisions I've made and that's the way we're doing things. todo is a bit of a mess with a todo section attached :grinning:. Took a deep look at the response from the rcAPI and added a dataDeepDive.txt which kind of destructures it and has notes on data type and what it represents. Saved a copy of the response for all countries to the project, allData.json. Will have to run some tests on this as some fields are not included for all countries, postalCode being only one I've found.