Project Update
Added jsonTest.js to backend. Revealed that the number of keys returned per country varies, 26 to 35. All countries with 35 keys have exactly the same 35 keys, so we're good. Added responseObject.js to backend. This will be the default object that I will send back to the frontend. It will have empty values that will be populated from the response I get back from rcAPI. Added a 'Steps Taken' section to my plan.md. Seems like a good idea to record a big picture of what I've done, where the commits are more detailed.