Project Update
Created arrayOfObjectsFromKeyValue.js in utils. The purpose is to take an object with keys, and use those keys to create new objects and place those new objects into an array. The original object's keys reference primitive values. For each new object pushed into the array, the primitive values are added to them using a key that is provided as an argument to the function. Renamed createResponseObject to populateResponseObject. Renamed responseObject to createResponseObject. createResponseObject is now a function that initializes a new responseObject. populateResponseObject takes the restcountries API response and the responseObject and populates the responseObject so it can be sent to the front end.