Project Update
Continuing to manually add tables and rows to my DynamicObjectRenderer. Added a TableRow component to play with. Fixed error cause by 'demonyms' field - this can be undefined. Error occurred because I checked if object was undefined after I checked if it was empty. Needed to do it the other way round! Added formatting for area in square kilometres. Fixed various issues around undefined data or arrays that only contain an empty string, empty or undefined objects. Updated TableRow component so it can handle arrays where the elements of the array are strings that can be output as commma separated values.