Project Update
Cleaned out todo.md to gain a bit more focus on what still needs to be done. Added Loading.js component to display when loading. Adjusted SearchForm transition off the screen. Added TableElementWithToolTip.js component, which accepts two string args and a boolean arg. headingText is the text that appears as the table heading. toolTipText is the tooltip text to be displayed on hover or touch. It renders a react icon that when moused over displays the tool tip text. The boolean arg, isHeading, determines whether the heading text is centered or not. Setting isHeading to true causes the heading text to be centered.