Project Update
Navbar is now rendered by BasciUI.js. Should probably get rid of App.js or move everything from BasicUI into App. I want my SearchForm component to transition off screen when I receive my data (perform my first search). Then I want it to appear with different styles in my Navbar component. The easiest way, I think, to do this is useRef and forwardRef. It won't be in the Navbar as such, but it will be absolutely positioned in that area of the screen. Added forwardRef to SearchForm component. Feels a bit hacky but seems to be working so far and quicker to learn/easier to implement than framer-motion at this point.