Project Update
Added react-toastify for notifications to user. If useAxios hook (fetches list of countries to display in our Select) cannot retrieve a list of countries (404), we notify the user of the error. Added isLoading prop to our Select component, so it now indicates that options are loading. If we turn off wifi or change the URL on our Node server, we get a 500 status with the same message. All good.