Project Update
Added toastify notification to catch block for our fetchCountryData request in SearchForm on front end. Tested backend error handling by altering the url we use to make requests to restcountries and by setting 'country,' the string received from the front end, to 'wales,' a value we know will cause restcountries to respond with 404. All errors are being handled on the frontend with a toastify notification and/or some conditional rendering logic.