Project Update
Fixed issue where currency was using the currency symbol as the key when rendering using map. Brunei has two currencies with the same symbol and so each item didn't have its own unqiue key. Set the key to be 3-letter currency code. Reordered my responseObject on backend so general category is rendered before geography. Decided not to use accordion for single column layout. Implemented essentially accordion logic for layouts with 2 or more columns, this addresses the fact that different info categories are different lengths when rendered. Now the user can click on a category to expand it.