Project Update
Added useEffect and useRef to SearchForm.js so we could implement AbortController on our axios fetch. Added 3 second timeout to our useAxios hook and our fetchCountryData request on the front end. Installed Cypress app and installed through npm and even edited some of the tests to test my own app, but couldn't get it working quickly so abandoned for now.